Move addon events into its own story file
