+++ examples/cra-kitchen-sink/src/stories/index.stories.js
@@ -1,28 +1,15 @@
 import React from 'react';
-import EventEmitter from 'eventemitter3';
 
 import { storiesOf } from '@storybook/react';
 import { setOptions } from '@storybook/addon-options';
 import { action } from '@storybook/addon-actions';
 import { withNotes, WithNotes } from '@storybook/addon-notes';
-import WithEvents from '@storybook/addon-events';
 import centered from '@storybook/addon-centered';
 import { withInfo } from '@storybook/addon-info';
 import { Button } from '@storybook/react/demo';
 
 import App from '../App';
-import Logger from './Logger';
 import Container from './Container';
-
-const EVENTS = {
-  TEST_EVENT_1: 'test-event-1',
-  TEST_EVENT_2: 'test-event-2',
-  TEST_EVENT_3: 'test-event-3',
-  TEST_EVENT_4: 'test-event-4',
-};
-
-const emitter = new EventEmitter();
-const emit = emitter.emit.bind(emitter);
 
 const InfoButton = () => (
   <span
@@ -101,60 +88,3 @@ storiesOf('App', module).add('full app', () => <App />);
 storiesOf('Some really long story kind description', module)
   .addDecorator(centered)
   .add('with text', () => <Button onClick={action('clicked')}>Hello Button</Button>);
-
-storiesOf('WithEvents', module)
-  .addDecorator(getStory => (
-    <WithEvents
-      emit={emit}
-      events={[
-        {
-          name: EVENTS.TEST_EVENT_1,
-          title: 'Test event 1',
-          payload: 0,
-        },
-        {
-          name: EVENTS.TEST_EVENT_2,
-          title: 'Test event 2',
-          payload: 'Test event 2',
-        },
-        {
-          name: EVENTS.TEST_EVENT_3,
-          title: 'Test event 3',
-          payload: {
-            string: 'value',
-            number: 123,
-            array: [1, 2, 3],
-            object: {
-              string: 'value',
-              number: 123,
-              array: [1, 2, 3],
-            },
-          },
-        },
-        {
-          name: EVENTS.TEST_EVENT_4,
-          title: 'Test event 4',
-          payload: [
-            {
-              string: 'value',
-              number: 123,
-              array: [1, 2, 3],
-            },
-            {
-              string: 'value',
-              number: 123,
-              array: [1, 2, 3],
-            },
-            {
-              string: 'value',
-              number: 123,
-              array: [1, 2, 3],
-            },
-          ],
-        },
-      ]}
-    >
-      {getStory()}
-    </WithEvents>
-  ))
-  .add('Logger', () => <Logger emitter={emitter} />);
